Parse ngModel.$modelValue to Duration object
If I set the ng-model to a valid Duration object and then try to change/update the ng-model through this directive, I get the error "TypeError: duration.get is not a function (angular-moment-duration.js:71). ngModel.$modelValue seems to be a String, not a Duration object. This PR makes Moment.js parse the $modelValue to a valid Duration object, which fixes the issue for me.